Cleanup resource and methods normalization
