+++ tests/spec/Normalizer/ResourcesNormalizerSpec.php
@@ -7,49 +7,23 @@ use Prophecy\Argument;
 
 class ResourcesNormalizerSpec extends ObjectBehavior
 {
-    function it_collects_resource_definitions()
-    {
-        $this->normalize([
-            'title' => 'API',
-            '/' => [],
-            '/users' => ['description' => '']
-        ])->shouldBeLike([
-            'title' => 'API',
-            'resources' => [
-                [
-                    'uri' => '/',
-                    'resources' => []
-                ],
-                [
-                    'uri' => '/users',
-                    'description' => '',
-                    'resources' => []
-                ]
-            ]
-        ]);
-    }
 
-    function it_normalizes_nested_resource_definitions()
+    function it_collects_methods()
     {
         $this->normalize([
-            '/users' => [
-                '/{id}' => [
-                    'description' => 'user details'
-                ]
-            ]
+            'uri' => '/users',
+            'test' => 'test',
+            'post' => [],
+            'get' => [],
+            'resources' => []
         ])->shouldBeLike([
-            'resources' => [
-                [
-                    'uri' => '/users',
-                    'resources' => [
-                        [
-                            'description' => 'user details',
-                            'uri' => '/{id}',
-                            'resources' => []
-                        ]
-                    ]
-                ],
-            ]
+            'uri' => '/users',
+            'test' => 'test',
+            'resources' => [],
+            'methods' => [
+                ['method' => 'post'],
+                ['method' => 'get']
+            ],
         ]);
     }
 }
